+++ framework/db-http/src/main/java/is/codion/framework/db/http/DefaultHttpEntityConnectionProvider.java
@@ -78,7 +78,8 @@ final class DefaultHttpEntityConnectionProvider extends AbstractEntityConnection
 							.port(port)
 							.securePort(securePort)
 							.user(user())
-							.clientType(clientType())
+							.clientType(clientType().orElseThrow(() ->
+											new IllegalStateException("clientType must be specified")))
 							.clientId(clientId())
 							.json(json)
 							.https(https)