+++ framework/db-rmi/src/main/java/is/codion/framework/db/rmi/DefaultRemoteEntityConnectionProvider.java
@@ -102,8 +102,9 @@ final class DefaultRemoteEntityConnectionProvider extends AbstractEntityConnecti
 											server().connect(ConnectionRequest.builder()
 															.user(user())
 															.clientId(clientId())
-															.clientType(clientType())
-															.clientVersion(clientVersion())
+															.clientType(clientType().orElseThrow(() ->
+																			new IllegalStateException("clientType must be specified")))
+															.clientVersion(clientVersion().orElse(null))
 															.parameter(REMOTE_CLIENT_DOMAIN_TYPE, domainType().name())
 															.build())));
 		}